back-end/script_synergy_backend/requirements.txt
asgiref==3.8.0
Django==4.1.13
django-cors-headers==4.3.1
djangorestframework==3.15.0
djongo==1.3.6
dnspython==2.6.1
pymongo==3.12.3
pytz==2024.1
sqlparse==0.2.4
pyjwt==2.8.0
boto3>=1.15.0